requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-core
langchain-community
langchain-groq
langsmith
groq
faiss-cpu
sentence-transformers
huggingface-hub
transformers
torch
pymupdf
pillow
pytesseract
pdf2image
numpy
scipy
requests
packaging
tqdm